+++ frontend/src/App.jsx
@@ -4,7 +4,7 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import MainContent from './components/MainContent';
 import MainNav from './components/MainNav';
 import Footer from './components/Footer';
-import History from './pages/History';
+import SeriesHistory from './pages/SeriesHistory';
 import Overview from './pages/Overview';
 import Analysis from './pages/Analysis';
 import BuildHistory from './pages/BuildHistory';
@@ -56,7 +56,7 @@ const App = () => {
                                 <BuildHistory />
                             </Route>
                             <Route path="/series/:seriesId/history">
-                                <History />
+                                <SeriesHistory />
                             </Route>
                             <Route path="/series/:seriesId/overview">
                                 <Overview />
