+++ frontend/src/pages/SeriesHistory.styles.js
